requirements.txt
Django==4.2.2
djangorestframework~=3.14.0
drf-yasg
psycopg2